Make the topical event document search reusable
This will allow us to reuse this code for World Location News, without
writing most of it again.
